pyproject.toml
[build-system]
requires = ["setuptools>=42.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dwarf_discord_bot"
version = "0.1.0"
description = "디스코드 봇 프로젝트"
authors = [
    {name = "개발자", email = "dev@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "discord.py>=2.3.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "aiohttp>=3.9.1",
    "urllib3>=2.0.7",
    "openai>=1.14.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]
dev = [
    "ruff>=0.3.0",
]

[tool.ruff]
target-version = "py39"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
]
fixable = ["ALL"]
ignore = []

[tool.ruff.isort]
known-first-party = ["cogs", "services", "utils"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto" 